Fix incorrect regex for amount conversion.
The regex in Ofx::createAmountFromStr included a "?" after the "mark"
whether it be decimal or comma.  That "?" made it match number strings
regardless of whether they had a "mark" to parse out, and grossly
changed their value.

This made amounts such as 1000 or -1000 convert to 10 or -10.  Removing
the conditional after the mark resolved this issue.  There are new tests
to show that this resolves the issue without breaking others.

Resolves #19