First test if dateString is set and not empty in createDateTimeFromStr
Some OFX files don't have all the date tags. 
So when the value of the date tag that doesn't exist and it is passed as a parameter to $dateString into the function "createDateTimeFromStr", this function returns "=> Failed to initialize DateTime for string:  ".
To bring more flexibility, I suggest testing first if dateString is set and is not empty before converting it into a string.